fix(metrics): calibration sampleSize is fractional (weighted effective-n), not int
summary.calibration[].sampleSize is Σ vindicator label weights (Phase 2b's
anti-gaming weighted effective-n) — fractional by design (e.g. 27.7 over 30d),
but the contract constrained it to .int(). A real 30d payload failed the
desktop's Zod parse of the whole detail → 'Couldn't load metrics. Retry'. 7d
(no calibration) validated, masking it. Relax to z.number().nonnegative(),
matching the sibling scoreBreakdown.calibrationMix.sampleSize (already float).
Regression test reproduces the exact 'Expected integer, received float'.

+++ packages/contracts/src/metrics/index.ts
@@ -72,7 +72,11 @@ export const TemplateMetricsSummary = z.object({
   calibration: z.array(z.object({
     source: z.enum(["executable", "grounding", "independent_review", "self_report"]),
     assumed: z.number(), measured: z.number().nullable(),
-    sampleSize: z.number().int().nonnegative(),
+    // Σ vindicator label weights — a weighted effective-n, fractional by design (a
+    // weak vindicator contributes <1). NOT an integer count; mirrors the sibling
+    // scoreBreakdown.calibrationMix.sampleSize (z.number()). The prior .int() rejected
+    // real 30d payloads (e.g. 27.7) and broke the whole metrics tab.
+    sampleSize: z.number().nonnegative(),
     state: z.enum(["measured", "insufficient", "unmeasurable"]),
   }).strict()),
   gateHealth: z.object({
